+++ source/isaaclab_tasks/isaaclab_tasks/core/lift/config/franka_soft/__init__.py
@@ -31,6 +31,29 @@ gym.register(
     },
 )
 
+
+gym.register(
+    id="Isaac-Lift-Cable-Franka",
+    entry_point="isaaclab.envs:ManagerBasedRLEnv",
+    disable_env_checker=True,
+    kwargs={
+        "env_cfg_entry_point": f"{__name__}.franka_cable_env_cfg:FrankaCableEnvCfg",
+        "rsl_rl_cfg_entry_point": f"{agents.__name__}.rsl_rl_ppo_cfg:FrankaCablePPORunnerCfg",
+    },
+)
+
+
+gym.register(
+    id="Isaac-Lift-Cable-Franka-Camera",
+    entry_point="isaaclab.envs:ManagerBasedRLEnv",
+    disable_env_checker=True,
+    kwargs={
+        "env_cfg_entry_point": f"{__name__}.franka_cable_env_cfg:FrankaCableCameraEnvCfg",
+        "rsl_rl_cfg_entry_point": f"{agents.__name__}.rsl_rl_ppo_cfg:FrankaCableCameraPPORunnerCfg",
+    },
+)
+
+
 gym.register(
     id="Isaac-Lift-Soft-Franka-Camera",
     entry_point="isaaclab.envs:ManagerBasedRLEnv",